rename inkstitch/ to lib/
You can't have a module and a package named the same thing.  PyInstaller wants
to import the main script as if it were a module, and this doesn't work unless
there's no directory of the same name with a __init__.py in it.
