+++ lib/elements/polyline.py
