+++ lib/elements/stroke.py
