+++ lib/extensions/palettes.py
