+++ lib/stitch_plan/stitch_plan.py
