+++ lib/stitch_plan/stop.py
