+++ lib/stitch_plan/ties.py
