+++ lib/stitch_plan/trim.py
