+++ lib/svg.py
