+++ lib/threads/palette.py
