+++ lib/utils/geometry.py
